Added react-dom, ts-loader and webpack as dependencies
React-dom is the browser component of react.js in recent versions, factored out as react.js core sees wider use beyond the confines of browsers.

WebPack is a package utility. We currently use it to package client-side scripts.

Ts-loader is a WebPack loader plugin to insert TypeScript -> JavaScript compilation in the package pipeline.